Extract meta custom sources to a separate file
